+++ HISTORY.md
@@ -6,9 +6,9 @@ History
 * Type changes.
   * `ReactComponentB` innards reworked. Use `ReactComponentB[P,S,B]` in place of `ReactComponentB[P]#B2[S]#B3[B]#B4[C]`.
   * `CompCtor`   → `ReactComponentC`
-  * `CompCtorP`  → `ReactComponentC.NeedProps`
-  * `CompCtorOP` → `ReactComponentC.OptionalProps`
-  * `CompCtorNP` → `ReactComponentC.NoProps`
+  * `CompCtorP`  → `ReactComponentC.ReqProps`
+  * `CompCtorOP` → `ReactComponentC.DefaultProps`
+  * `CompCtorNP` → `ReactComponentC.ConstProps`
 
 ### 0.4.1 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.4.0...v0.4.1))